Initialise new rails application with tailwind
The new rails app is Initialise using the following command and flags

rails new search_track --skip-action-mailbox --skip-action-mailer --css=tailwind --skip-test --database=postgresql

We skip both mailbox and mailer as we intend to not use them.
Tailwind is used for css while the defauls minitest is skipped in favor of
rspec that will be setup later and postgresql is used as the db.