Add searches controller to display analytics
This commit adds the index action that fetches top
and recent queries from the model so as to be
rendered to the view.

The commit also adds request tests to verify that
the search analytics are fetched for the current
user's IP address to be rendered properly.

Routes is updated to access this controller
including routes tests.